backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic (compatible versions)
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# AI/ML
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
numpy==1.24.3
scipy==1.11.4

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.16.3
google-auth==2.23.4

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP Requests
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3

# File Processing
Pillow==10.1.0
python-magic==0.4.27

# Audio/Video
moviepy==1.0.3

# Utilities
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Async
asyncio-mqtt==0.16.1
celery==5.3.4
redis==5.0.1

# Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
stripe>=5.0.0
